fix(homepage,homepage,me,api): coderabbit reviews
